+++ index.js
@@ -25,8 +25,7 @@ const errorHandler = (error) => {
   }
   const address = server.address();
 
-  const bind =
-    typeof address === "string" ? `pipe ${address}` : `port: ${port}`;
+  const bind = typeof address === "string" ? `pipe ${address}` : `port: ${port}`;
   switch (error.code) {
     case "EACCES":
       console.error(`${bind} requires elevated privileges.`);
@@ -44,8 +43,7 @@ const errorHandler = (error) => {
 server.on("error", errorHandler);
 server.on("listening", () => {
   const address = server.address();
-  const bind =
-    typeof address === "string" ? `pipe ${address}` : `port: ${port}`;
+  const bind = typeof address === "string" ? `pipe ${address}` : `port: ${port}`;
   console.log(`Listening on ${bind}`);
 });
