+++ package.json
@@ -20,7 +20,7 @@
     "jsonwebtoken": "^8.5.1",
     "lodash": "^4.17.21",
     "mailgun": "^0.5.0",
-    "mailgun-js": "^0.22.0",
+    "mailgun-js": "^0.6.7",
     "mongoose": "^5.12.14"
   },
   "devDependencies": {
